_WidgetTree type change
Fixed rending bug which occurred when anscestor and descendent elements
are re-rendered at the same time.

RenderEngine._component_tree type errors fixed.

use_state will not re-render if state unchanged after update.

use_state track updates with use_state_setted.

use_async window close bugfix.

Delete prototype WidgetElement.

Deprecate Table.
